Use keychain instead of .credentials.env
